+++ packages/xchain-util/src/index.ts
@@ -4,3 +4,5 @@ export { default as bn, isValidBN, bnOrZero, validBNOrZero, formatBN, formatBNCu
 export * from './string'
 export * from './types'
 export * from './ninerealms'
+export * from './crypto-amount'
+export * from './cached-value'